refactor(audit): split extension/manifest CI config (#5645)
