+++ src/core/extension/manifest.rs
@@ -22,6 +22,10 @@ pub use super::manifest_config::{
     TraceBrowserMetricAliasConfig, TraceBrowserSummaryAliasConfig, TraceConfig,
     VersionPatternConfig,
 };
+mod ci_config;
+pub use ci_config::{
+    CiCapability, CiJobFidelity, CiJobMapping, CiJobSpec, CiLocalContext, CiProfileSpec,
+};
 pub use super::manifest_deploy_config::DeployArchiveInstallPolicy;
 pub use super::manifest_sidecar::{StructuredSidecarContract, StructuredSidecarDeclaration};
 pub use super::manifest_test_config::{TestPassthroughFilter, TestPassthroughFilterStrategy};
@@ -389,69 +393,6 @@ pub struct ComponentEnvConfig {
     /// Script path relative to the extension directory.
     /// Runs from the component root and emits runtime metadata as JSON.
     pub detect_script: String,
-}
-
-/// CI reproduction profiles declared by an extension.
-///
-/// Core treats these as explicit local-reproduction contracts. Best-effort
-/// workflow discovery can inventory CI files, but runnable equivalence comes
-/// from these declarations.
-#[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq, Eq)]
-pub struct CiCapability {
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub profiles: BTreeMap<String, CiProfileSpec>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub jobs: BTreeMap<String, CiJobSpec>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq, Eq)]
-pub struct CiProfileSpec {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub label: Option<String>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub jobs: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "kebab-case")]
-#[derive(Default)]
-pub enum CiJobFidelity {
-    LocalEquivalent,
-    LocalPartial,
-    RemoteOnly,
-    #[default]
-    Unknown,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq, Eq)]
-pub struct CiJobSpec {
-    #[serde(flatten)]
-    pub mapping: CiJobMapping,
-    pub command: String,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub args: Vec<String>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub env: BTreeMap<String, String>,
-    #[serde(flatten)]
-    pub local_context: CiLocalContext,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq, Eq)]
-pub struct CiJobMapping {
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub check_names: Vec<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub provider: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub workflow: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq, Eq)]
-pub struct CiLocalContext {
-    #[serde(default)]
-    pub fidelity: CiJobFidelity,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub limitations: Vec<String>,
 }
 
 // ============================================================================
